+++ routes/main.py
@@ -65,6 +65,17 @@ def parse_args(args=None):
         action="store_true",
         help="Open HTML output file in browser",
     )
+    viz_parser.add_argument(
+        "--global",
+        dest="_global",
+        action="store_true",
+        help="Expiremental: enable global security configuration detection",
+    )
+    viz_parser.add_argument(
+        "--interprocedural",
+        action="store_true",
+        help="Expiremental: enable interprocedural security configuration detection",
+    )
 
     return p.parse_args(args=args)