+++ tests/test_commands/test_viz.py
@@ -214,7 +214,12 @@ def test_viz_basic(data, expected):
     output = make_file("", "test_output")
     template = make_file(const.DEFAULT_TEMPLATE_KEY, "test_template")
     args = argparse.Namespace(
-        input=input, output=output, template=template, browser=False
+        input=input,
+        output=output,
+        template=template,
+        browser=False,
+        interprocedural=True,
+        _global=True,
     )
 
     code = commands.viz.main(args)
@@ -240,7 +245,12 @@ def test_viz_multiple_root():
     output = make_file("", "test_output")
     template = make_file(const.DEFAULT_TEMPLATE_KEY, "test_template")
     args = argparse.Namespace(
-        input=input, output=output, template=template, browser=False
+        input=input,
+        output=output,
+        template=template,
+        browser=False,
+        interprocedural=True,
+        _global=True,
     )
 
     with pytest.raises(ValueError):